+++ ext_tables.php
@@ -14,12 +14,3 @@ if (TYPO3_MODE === 'BE'
         'Tx_Phpunit_BackEnd_Ajax->ajaxBroker'
     );
 }
-
-$GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['GLOBAL']['cliKeys']['phpunit'] = [
-    'EXT:phpunit/Scripts/ManualCliTestRunner.php',
-    '_CLI_phpunit',
-];
-$GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['GLOBAL']['cliKeys']['phpunit_ide_testrunner'] = [
-    'EXT:phpunit/Scripts/IdeTestRunner.php',
-    '_CLI_phpunit',
-];